+++ CHANGELOG.md
@@ -6,11 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `*_tt` functions to `TokenParser`
-- `next_macro_rules_tt()` to `TokenParser`
+- `*_tt` functions to `TokenParser`.
+- `next_macro_rules_tt()` to `TokenParser`.
+- `TokenTreeLiteral` with `is_string()` and `string()`.
+
+### Changed
+- **Breaking Change** Sealed extension traits.
 
 ### Fixed
-- `assert_expansion!` was failing on nightly
+- `assert_expansion!` was failing on nightly.
 
 ## [0.8.0] - 2023-05-14
 ### Fixed