+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [0.9.0] - 2024-03-13
 ### Added
 - `*_tt` functions to `TokenParser`.
 - `next_macro_rules_tt()` to `TokenParser`.
@@ -70,7 +72,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - **Breaking Change** `TokenTreeExt::<token>()` now return references
 
-[unreleased]: https://github.com/ModProg/proc-macro-utils/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/ModProg/proc-macro-utils/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/ModProg/proc-macro-utils/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ModProg/proc-macro-utils/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/ModProg/proc-macro-utils/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ModProg/proc-macro-utils/compare/v0.5.2...v0.6.0